Move autodispose sample to autodispose2.* package
