+++ sample/gradle.properties
@@ -1,3 +1,3 @@
-autodispose.typesWithScope=com.uber.autodispose.sample.CustomScope
+autodispose.typesWithScope=autodispose2.sample.CustomScope
 autodispose.lenient=false
 kapt.includeCompileClasspath=false
