+++ sample/src/main/AndroidManifest.xml
@@ -29,22 +29,22 @@
       android:fullBackupContent="@null"
       tools:ignore="GoogleAppIndexingWarning">
     <activity
-        android:name=".JavaActivity"
+        android:name="autodispose2.sample.JavaActivity"
         android:label="@string/app_name"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar"/>
     <activity
-        android:name=".KotlinActivity"
+        android:name="autodispose2.sample.KotlinActivity"
         android:label="@string/app_name_kotlin"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar"/>
     <activity
-        android:name=".ArchComponentActivity"
+        android:name="autodispose2.sample.ArchComponentActivity"
         android:label="@string/title_activity_arch_component"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar"/>
     <activity
-        android:name=".DisposingViewModelActivity"
+        android:name="autodispose2.sample.DisposingViewModelActivity"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar"/>
     <activity
-        android:name=".HomeActivity"
+        android:name="autodispose2.sample.HomeActivity"
         android:label="@string/app_name"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar">
